+++ packages/fighting-design/skeleton/src/props.ts
@@ -3,48 +3,41 @@ import type { SkeletonSizeType } from './interface'
 
 export const Props = {
   /**
-   * 是否带圆角的
+   * @zh 是否带圆角的
    */
-  rounded: {
+  round: {
     type: Boolean,
     default: (): boolean => false
   },
   /**
-   * 是否展示波浪动画
+   * @zh 是否展示波浪动画
    */
   animated: {
     type: Boolean,
     default: (): boolean => false
   },
   /**
-   * 圆形占位图
-   */
-  circled: {
-    type: Boolean,
-    default: (): boolean => false
-  },
-  /**
-   * 渲染多行占位图
+   * @zh 渲染多行占位图
    */
   rows: {
     type: Number,
     default: (): number => 1
   },
   /**
-   * 为 true 时，显示占位图。false 展示子组件
+   * @zh 为 true 时，显示占位图。false 展示子组件
    */
   loading: {
     type: Boolean,
     default: (): boolean => false
   },
   /**
-   * 设置尺寸，提供三个属性值，默认 default
+   * @zh 设置尺寸，提供三个属性值，默认 default
    */
   size: {
     type: String as PropType<SkeletonSizeType>,
-    default: (): SkeletonSizeType => 'default',
+    default: (): SkeletonSizeType => 'middle',
     validator: (val: SkeletonSizeType): boolean => {
-      return (['default', 'small', 'large'] as const).includes(val)
+      return (['large', 'middle', 'small', 'mini'] as const).includes(val)
     }
   }
 } as const