+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - `f-breadcrumb` 组件 `separator-color` 参数改为 `icon-color`
 - `f-breadcrumb` 组件 `item-color` 参数改为 `color`
 - `f-breadcrumb-item` 组件新增 `color` `icon-color` `separator` 配置项
+- `f-skeleton` 组件废除 `circled` 配置项
 
 ## 0.9.0-alpha.3 (2022-11-18)